+++ contrib/localnet/scripts/start-zetaclientd.sh
@@ -69,7 +69,7 @@ RELAYER_KEY_PATH="$HOME/.zetacored/relayer-keys"
 mkdir -p "${RELAYER_KEY_PATH}"
 
 mkdir -p "$HOME/.tss/"
-zetae2e get-zetaclient-bootstrap > "$HOME/.tss/address_book.seed"
+zetae2e local get-zetaclient-bootstrap > "$HOME/.tss/address_book.seed"
 
 MYIP=$(/sbin/ip -o -4 addr list eth0 | awk '{print $4}' | cut -d/ -f1)
